requirements/base.txt
python-dateutil==2.7.3
pytzdata==2018.5
redis==2.10.6
six==1.11.0
typed-ast==1.1.0
ujson==1.35
uvloop==0.11.2
vibora==0.0.6
whitenoise==4.1
wrapt==1.10.11
python-dotenv==0.8.2
aioredis==1.1.0